fixes with review
